Clean up the class stucture and align against stores
