+++ classes/local/form/settings_form.php
@@ -23,7 +23,7 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace tool_objectfs\form;
+namespace tool_objectfs\local\form;
 
 defined('MOODLE_INTERNAL') || die();
