+++ classes/local/object_manipulator/logger.php
@@ -22,7 +22,7 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace tool_objectfs\object_manipulator;
+namespace tool_objectfs\local\object_manipulator;
 
 defined('MOODLE_INTERNAL') || die();
