+++ classes/local/object_manipulator/recoverer.php
@@ -15,7 +15,7 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Pulls files from remote storage if they meet the configured criterea.
+ * Recovers objects that are in the error state if it can.
  *
  * @package   tool_objectfs
  * @author    Kenneth Hendricks <kennethhendricks@catalyst-au.net>
@@ -23,31 +23,23 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace tool_objectfs\object_manipulator;
+namespace tool_objectfs\local\object_manipulator;
 
 defined('MOODLE_INTERNAL') || die();
 
 require_once($CFG->dirroot . '/admin/tool/objectfs/lib.php');
 
-class puller extends manipulator {
+class recoverer extends manipulator {
 
     /**
-     * Size threshold for pulling files from remote in bytes.
+     * recoverer constructor.
      *
-     * @var int
-     */
-    private $sizethreshold;
-
-    /**
-     * Puller constructor.
-     *
-     * @param object_client $client object client
-     * @param object_file_system $filesystem object file system
-     * @param object $config objectfs config.
+     * @param sss_client $client S3 client
+     * @param object_file_system $filesystem S3 file system
+     * @param object $config sssfs config.
      */
     public function __construct($filesystem, $config, $logger) {
         parent::__construct($filesystem, $config);
-        $this->sizethreshold = $config->sizethreshold;
 
         $this->logger = $logger;
         // Inject our logger into the filesystem.
@@ -55,7 +47,7 @@ class puller extends manipulator {
     }
 
     protected function get_query_name() {
-        return 'get_pull_candidates';
+        return 'get_recover_candidates';
     }
 
     protected function get_candidates_sql() {
@@ -64,25 +56,23 @@ class puller extends manipulator {
                        MAX(f.filesize) AS filesize
                   FROM {files} f
              LEFT JOIN {tool_objectfs_objects} o ON f.contenthash = o.contenthash
+                 WHERE o.location = ?
               GROUP BY f.contenthash,
                        f.filesize,
-                       o.location
-                HAVING MAX(f.filesize) <= ?
-                       AND (o.location = ?)';
+                       o.location';
 
         return $sql;
     }
 
     protected function get_candidates_sql_params() {
-        $params = array($this->sizethreshold, OBJECT_LOCATION_EXTERNAL);
+        $params = array(OBJECT_LOCATION_ERROR);
 
         return $params;
     }
 
     protected function manipulate_object($objectrecord) {
-        $newlocation = $this->filesystem->copy_object_from_external_to_local_by_hash($objectrecord->contenthash, $objectrecord->filesize);
+        $newlocation = $this->filesystem->get_object_location_from_hash($objectrecord->contenthash, $objectrecord->filesize);
         return $newlocation;
     }
+
 }
-
-
