+++ classes/local/report/location_report_builder.php
@@ -23,7 +23,7 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace tool_objectfs\report;
+namespace tool_objectfs\local\report;
 
 defined('MOODLE_INTERNAL') || die();
