+++ classes/local/store/object_client_base.php
@@ -23,7 +23,7 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace tool_objectfs\client;
+namespace tool_objectfs\local\store;
 
 defined('MOODLE_INTERNAL') || die();
