+++ classes/s3_file_system.php
@@ -30,53 +30,8 @@ namespace tool_objectfs;
 
 defined('MOODLE_INTERNAL') || die();
 
-use tool_objectfs\client\s3_client;
+use tool_objectfs\local\store\s3\file_system;
 
-require_once($CFG->dirroot . '/admin/tool/objectfs/lib.php');
+class s3_file_system extends file_system {
 
-class s3_file_system extends object_file_system {
-
-    protected function initialise_external_client($config) {
-        $s3client = new s3_client($config);
-
-        return $s3client;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function readfile(\stored_file $file) {
-        $path = $this->get_remote_path_from_storedfile($file);
-
-        $this->get_logger()->start_timing();
-        if ($path == $this->get_external_client()->get_fullpath_from_hash($file->get_contenthash())) {
-            // There is an issue using core readfile_allow_large() for the big (more than 1G) files from s3.
-            $success = readfile($path);
-        } else {
-            $success = readfile_allow_large($path, $file->get_filesize());
-        }
-        $this->get_logger()->end_timing();
-        $this->get_logger()->log_object_read('readfile', $path, $file->get_filesize());
-
-        if (!$success) {
-            update_object_record($file->get_contenthash(), OBJECT_LOCATION_ERROR);
-        }
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function copy_from_local_to_external($contenthash) {
-        $localpath = $this->get_local_path_from_hash($contenthash);
-
-        try {
-            $this->get_external_client()->upload_to_s3($localpath, $contenthash);
-            return true;
-        } catch (\Exception $e) {
-            $this->get_logger()->error_log(
-                'ERROR: copy ' . $localpath . ' to ' . $this->get_external_path_from_hash($contenthash) . ': ' . $e->getMessage()
-            );
-            return false;
-        }
-    }
 }
