+++ tests/classes/test_client.php
@@ -18,8 +18,7 @@ namespace tool_objectfs\tests;
 
 defined('MOODLE_INTERNAL') || die();
 
-use tool_objectfs\object_file_system;
-use tool_objectfs\client\object_client_base;
+use tool_objectfs\local\store\object_client_base;
 
 class test_client extends object_client_base {
